Modified 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,7 +10,7 @@ followed by a new line
 int main(void)
 {
 int i;
-unsigned long int a, b, c;
+unsigned long long int a, b, c;
 
 a = 1;
 b = 2;
